loading/lazyload/move-element-and-scroll.tentative.html

<!DOCTYPE html>
<head>
  <title>Images with loading='lazy' load being moved to another document
         and then scrolled to</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="common.js"></script>
</head>

<!--
Marked as tentative until https://github.com/whatwg/html/pull/3752 is landed.
-->

<body>
  <img id="in_viewport" src='resources/image.png?in_viewport'>
  <div style="height:1000vh"></div>
  <div id="below_viewport_div"></div>
  <img id="below_viewport" src='resources/image.png?below_viewport' loading="lazy">

  <script>
    const in_viewport_element = document.getElementById("in_viewport");
    const below_viewport_element = document.getElementById("below_viewport");
    const below_viewport_div = document.getElementById("below_viewport_div");

    async_test(function(t) {
      below_viewport_element.onload =
        t.unreached_func("The below viewport image should not load");
      t.step_timeout(t.step_func_done(() => {
        assert_false(is_image_fully_loaded(below_viewport_element,
                                           in_viewport_element));
      }), 1000);
      const iframe = document.createElement('iframe');
      iframe.setAttribute("display", "none");
      iframe.srcdoc = "<body></body>";
      iframe.onload = () => {
        var adopted_img = iframe.contentDocument.adoptNode(below_viewport_element);
        iframe.contentDocument.body.appendChild(adopted_img);
        below_viewport_div.scrollIntoView();
      };
      document.body.insertBefore(iframe, in_viewport_element);
    }, "Test that <img> below viewport is not loaded when moved to another " +
       "document and then scrolled to");
</script>
</body>
